requirements.txt
absl-py==1.0.0
antlr4-python3-runtime==4.8
anyio==3.6.2
asttokens==2.2.1
astunparse==1.6.3
autocorrect==2.6.0
backcall==0.2.0
beautifulsoup4==4.10.0
cachetools==4.2.4
camelot-py==0.10.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
comm==0.1.2
compressed-rtf==1.0.6
cryptography==36.0.1
cycler==0.10.0
Cython==0.29.24
debugpy==1.6.5
decorator==5.1.1
distlib==0.3.7
distro==1.6.0
easyocr==1.4.1
ebcdic==1.1.1
editdistance==0.6.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
extract-msg==0.28.7
fastapi==0.89.1
filelock==3.12.2
Flask==2.0.2
flask-paginate==2022.1.8
flatbuffers==2.0
gast==0.4.0
ghostscript==0.7
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.1
h11==0.14.0
h5py==3.6.0
idna==3.3
imageio==2.9.0
IMAPClient==2.1.0
imutils==0.5.4
independentsoft.msg==1.6
iopath==0.1.9
ipykernel==6.20.2
ipython==8.8.0
itsdangerous==2.0.1
jedi==0.18.2
Jinja2==3.0.2
joblib==1.1.0
jupyter_client==7.4.9
jupyter_core==5.1.3
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
layoutparser==0.3.2
ldap3==2.9.1
libclang==12.0.0
lmdb==1.2.1
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
networkx==2.6.3
nltk==3.6.6
numpy==1.21.2
oauthlib==3.1.1
olefile==0.46
opencv-contrib-python==4.5.4.58
opencv-python==4.5.4.58
opencv-python-headless==4.5.4.58
openpyxl==3.0.9
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.4
parso==0.8.3
path==16.2.0
pdf2image==1.16.0
pdfkit==1.0.0
pdfminer.six==20211012
pdfplumber==0.5.28
pdftopng==0.2.3
pickleshare==0.7.5
PIL-Tools==1.1.0
Pillow==8.2.0
plac==1.3.5
platformdirs==3.9.1
plotly==5.5.0
portalocker==2.3.2
prompt-toolkit==3.0.36
protobuf==3.19.1
psutil==5.9.4
psycopg2==2.9.3
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.11.0
pydantic==1.10.4
pydirectory==0.1.15
pydot==1.4.2
Pygments==2.14.0
pymongo==4.0.1
PyMuPDF==1.19.6
pyparsing==2.4.7
PyPDF2==1.26.0
pypiwin32==223
pytesseract==0.3.9
python-bidi==0.4.2
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.2.0
pywin32==302
PyYAML==6.0
pyzmq==25.0.0
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.3
scikit-learn==1.0
scipy==1.7.1
seaborn==0.11.2
six==1.16.0
sklearn==0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.2.1
stack-data==0.6.2
starlette==0.22.0
tabula-py==2.3.0
tabulate==0.8.9
tenacity==8.0.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
tesseract==0.1.3
threadpoolctl==3.0.0
tifffile==2021.11.2
tk==0.1.0
torch==1.10.0
torchvision==0.11.1
tornado==6.2
tqdm==4.62.3
traitlets==5.8.1
typing_extensions==4.4.0
tzdata==2021.5
tzlocal==4.1
Unidecode==1.3.4
urllib3==1.26.7
uvicorn==0.20.0
virtualenv==20.24.2
Wand==0.6.7
wcwidth==0.2.6
Werkzeug==2.0.2
wkhtmltopdf==0.2
wordninja==2.0.0
wrapt==1.13.3
yacs==0.1.8
